requirements.txt
django
django-cors-headers
pillow
djangorestframework
djangorestframework-simplejwt
requests